pyproject.toml
[project]
name = "llm-redactor"
version = "0.1.0"
description = "Privacy-preserving shim for outbound LLM requests"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Jay Lux Ferro", email = "jay@sperixlabs.org" },
]
dependencies = [
    "httpx>=0.27",
    "mcp>=1.0",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "pyyaml>=6.0",
    "presidio-analyzer>=2.2",
    "presidio-anonymizer>=2.2",
    "spacy>=3.7",
    "typer>=0.15",
    "rich>=13.0",
    "tiktoken>=0.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.35",
    "ruff>=0.8",
]

[project.scripts]
llm-redactor = "llm_redactor.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_redactor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
